+++ catalog/catalog_domain/src/main/kotlin/com/marmatsan/dev/catalog_domain/model/Plant.kt
@@ -22,3 +22,9 @@ enum class PlantSize {
     LARGE,
     EXTRA_LARGE
 }
+
+object PlantDataConstraints {
+    const val PLANT_NAME_MAX_LENGTH = 100
+    const val WATER_AMOUNT_MAX_LENGTH = 4
+    const val DESCRIPTION_MAX_LENGTH = 5000
+}
