--- core/core_domain/src/main/kotlin/com/marmatsan/dev/core_domain/catalog/PlantDataConstraints.kt
@@ -1,7 +1,0 @@
-package com.marmatsan.dev.core_domain.catalog
-
-object PlantDataConstraints {
-    const val PLANT_NAME_MAX_LENGTH = 100
-    const val WATER_AMOUNT_MAX_LENGTH = 4
-    const val DESCRIPTION_MAX_LENGTH = 5000
-}
